Add simple TODO app
The JQuery version is still a quick hack (reasonably good as I believe).

The Angular version has offered a very nice way to look at the modal-view
relationship. The next step should be to imitate this pattern in plain JQuery.

At this level of complexity, I've lost my interest to do it in plain JS (with
DOM API.....hrrrr)